# backend/openapi.yml
openapi: 3.0.3
info:
  title: Tickets API
  version: 1.0.0
  description: REST API для управления билетами, людьми, событиями и площадками.
servers:
  - url: /
...
paths:
...
  /add_event:
    post:
      tags: [Events]
      summary: Создать событие
      operationId: addEvent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Event' }
      responses:
        '200':
          description: Создано (data может содержать созданный Event)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
        '400':
          description: Ошибка
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /get_events:
    get:
      tags: [Events]
      summary: Получить список событий
      operationId: getEvents
      responses:
        '200':
          description: Успех (data = EventsEnvelope)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
        '400':
          description: Ошибка
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /add_person:
    post:
      tags: [Persons]
      summary: Создать человека
      operationId: addPerson
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Person' }
      responses:
        '200':
          description: Создано (data может содержать созданного Person)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
        '400':
          description: Ошибка
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /get_persons:
    get:
      tags: [Persons]
      summary: Получить список людей
      operationId: getPersons
      responses:
        '200':
          description: Успех (data = PersonEnvelope)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
        '400':
          description: Ошибка
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
components:
  schemas:
    AbstractResponse:
      type: object
      properties:
        title: { type: string, nullable: true }
        message: { type: string, nullable: true }
        status: { type: string, nullable: true, example: "OK" }
        data: { nullable: true, description: "Произвольные данные ответа" }
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/AbstractResponse'
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/AbstractResponse'
...
    Person:
      type: object
      properties:
        id: { type: integer, format: int32, nullable: true }
        passportID: { type: string, nullable: true }
        weight: { type: number, format: float, nullable: true }
        nationality:
          type: string
          nullable: true
          enum: [ GERMANY, INDIA, THAILAND, SOUTH_KOREA, JAPAN ]
        hairColor:
          type: string
          nullable: true
          enum: [ GREEN, RED, ORANGE, WHITE, BROWN ]
        eyeColor:
          type: string
          nullable: true
          enum: [ GREEN, RED, ORANGE, WHITE, BROWN ]
        location:
          type: object
          nullable: true
          properties:
            x: { type: integer, format: int32 }
            y: { type: number,  format: float }
            z: { type: number,  format: float }
          required: [ x, y, z ]
    Event:
      type: object
      properties:
        id: { type: integer, format: int32, nullable: true }
        name: { type: string }
        ticketsCount: { type: integer, format: int32, nullable: true }
        eventType:
          type: string
          nullable: true
          enum: [ CONCERT, FOOTBALL, BASEBALL, BASKETBALL, OPERA ]
...