# front/src/api/models/error-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tickets API
 * REST API для управления билетами, людьми, событиями и площадками.
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

// May contain unused imports in some cases
// @ts-ignore
import type { AbstractResponse } from './abstract-response';

/**
 * @type ErrorResponse
 * @export
 */
export type ErrorResponse = AbstractResponse;