# front/src/api/models/event.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tickets API
 * REST API для управления билетами, людьми, событиями и площадками.
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface Event
 */
export interface Event {
  /**
   *
   * @type {number}
   * @memberof Event
   */
  id?: number | null;
  /**
   *
   * @type {string}
   * @memberof Event
   */
  name?: string;
  /**
   *
   * @type {number}
   * @memberof Event
   */
  ticketsCount?: number | null;
  /**
   *
   * @type {string}
   * @memberof Event
   */
  eventType?: EventEventTypeEnum | null;
}

export const EventEventTypeEnum = {
  CONCERT: 'CONCERT',
  FOOTBALL: 'FOOTBALL',
  BASEBALL: 'BASEBALL',
  BASKETBALL: 'BASKETBALL',
  OPERA: 'OPERA',
} as const;

export type EventEventTypeEnum = (typeof EventEventTypeEnum)[keyof typeof EventEventTypeEnum];