# front/src/api/models/person.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tickets API
 * REST API для управления билетами, людьми, событиями и площадками.
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

// May contain unused imports in some cases
// @ts-ignore
import type { PersonLocation } from './person-location';

/**
 *
 * @export
 * @interface Person
 */
export interface Person {
  /**
   *
   * @type {number}
   * @memberof Person
   */
  id?: number | null;
  /**
   *
   * @type {string}
   * @memberof Person
   */
  passportID?: string | null;
  /**
   *
   * @type {number}
   * @memberof Person
   */
  weight?: number | null;
  /**
   *
   * @type {string}
   * @memberof Person
   */
  nationality?: PersonNationalityEnum | null;
  /**
   *
   * @type {string}
   * @memberof Person
   */
  hairColor?: PersonHairColorEnum | null;
  /**
   *
   * @type {string}
   * @memberof Person
   */
  eyeColor?: PersonEyeColorEnum | null;
  /**
   *
   * @type {PersonLocation}
   * @memberof Person
   */
  location?: PersonLocation | null;
}

export const PersonNationalityEnum = {
  GERMANY: 'GERMANY',
  INDIA: 'INDIA',
  THAILAND: 'THAILAND',
  SOUTH_KOREA: 'SOUTH_KOREA',
  JAPAN: 'JAPAN',
} as const;

export type PersonNationalityEnum =
  (typeof PersonNationalityEnum)[keyof typeof PersonNationalityEnum];
export const PersonHairColorEnum = {
  GREEN: 'GREEN',
  RED: 'RED',
  ORANGE: 'ORANGE',
  WHITE: 'WHITE',
  BROWN: 'BROWN',
} as const;

export type PersonHairColorEnum = (typeof PersonHairColorEnum)[keyof typeof PersonHairColorEnum];
export const PersonEyeColorEnum = {
  GREEN: 'GREEN',
  RED: 'RED',
  ORANGE: 'ORANGE',
  WHITE: 'WHITE',
  BROWN: 'BROWN',
} as const;

export type PersonEyeColorEnum = (typeof PersonEyeColorEnum)[keyof typeof PersonEyeColorEnum];